feat: メンバーテーブルをコンポーネント化する #119
