+++ frontend/src/app/routes/app/management/groupNew.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useMemo, useState } from "react";
 import { useForm } from "react-hook-form";
 import Select, { type MultiValue, type ActionMeta } from "react-select";
+import { MemberTable } from "@/features/management/components/memberTable/memberTable";
 import { type GroupsOutletContext } from "@/features/management/layouts/groupShell/groupShell";
 import {
   type GroupRequestType,
@@ -207,57 +208,11 @@ const GroupNew = () => {
           </div>
         </div>
 
-        <div className={styles.memberTableContainer}>
-          <table className={styles.memberTable}>
-            <thead>
-              <tr>
-                <th>ユーザID</th>
-                <th>名前</th>
-                <th>学年</th>
-                <th>参加状況</th>
-              </tr>
-            </thead>
-            <tbody>
-              {data ? (
-                data.map((member, index) => {
-                  if (searchQuery) {
-                    const lowerQuery = searchQuery.toLowerCase();
-                    const lowerName = member.name.toLowerCase();
-                    const lowerId = member.showUserId.toLowerCase();
-
-                    if (
-                      !lowerName.includes(lowerQuery) &&
-                      !lowerId.includes(lowerQuery)
-                    ) {
-                      return null;
-                    }
-                  }
-
-                  return (
-                    <tr key={member.userId}>
-                      <td>{member.showUserId}</td>
-                      <td>{member.name}</td>
-                      <td>{member.grade}年</td>
-                      <td>
-                        <input
-                          type="checkbox"
-                          {...register(`members.${index}.isJoined`)}
-                        />
-                        <input
-                          type="hidden"
-                          value={member.userId}
-                          {...register(`members.${index}.userId`)}
-                        />
-                      </td>
-                    </tr>
-                  );
-                })
-              ) : (
-                <></>
-              )}
-            </tbody>
-          </table>
-        </div>
+        <MemberTable
+          data={data}
+          searchQuery={searchQuery}
+          register={register}
+        />
       </div>
     </form>
   );
