+++ frontend/src/features/management/components/memberTable/memberTable.module.css
@@ -1,0 +1,45 @@
+.memberTableContainer {
+  flex: 1;
+  min-height: 0;
+}
+
+.memberTable {
+  width: 100%;
+  border-collapse: separate;
+  border-spacing: 0;
+  background: #fff;
+  table-layout: fixed;
+  border: 1px solid #e2e8f0;
+  border-radius: 10px;
+}
+
+.memberTable thead th:nth-child(3),
+.memberTable thead th:nth-child(4),
+.memberTable tbody td:nth-child(3),
+.memberTable tbody td:nth-child(4) {
+  text-align: center;
+  width: 5.5rem;
+  white-space: nowrap;
+}
+
+.memberTable thead th {
+  position: sticky;
+  top: 0;
+  z-index: 1;
+  background: rgb(50, 85, 150);
+  color: #fff;
+  text-align: left;
+  padding: 0.55rem;
+  font-size: 0.9rem;
+  border-bottom: 1px solid rgb(56, 94, 164);
+}
+
+.memberTable tbody td {
+  padding: 0.75rem 0.9rem;
+  border-bottom: 1px solid #eef2f7;
+  font-size: 0.9rem;
+}
+
+.memberTable tbody tr:hover td {
+  background: #f8fafc;
+}
